Moved SyntaxKind map type and delegator object to a common file so I dont have to edit more files when adding additional support
